[rpc][api] Replace Product.filter_versions -> Version.filter and
	Product.add_version with Version.create
