+++ docs/source/modules/tcms.xmlrpc.api.version.rst
@@ -1,0 +1,7 @@
+tcms\.xmlrpc\.api\.version module
+=================================
+
+.. automodule:: tcms.xmlrpc.api.version
+    :members:
+    :undoc-members:
+    :show-inheritance:
